requirements.txt
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
tzdata==2025.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
